PI-3163 Only show invalid number stats in header
Also increase timeout + add duplicates page + separate routes files
