+++ integration_tests/e2e/data-quality.cy.ts
@@ -6,11 +6,12 @@ context('Data quality page', () => {
     cy.visit('/')
     Page.verifyOnPage(IndexPage)
 
-    cy.get('#notifications').should('contain.text', '1,910')
+    cy.get('#notifications').should('contain.text', '99+')
     cy.get('#notifications').click()
     cy.url().should('contain', '/data-quality/invalid?provider=N56')
-    cy.get('.moj-sub-navigation__link').first().should('contain.text', 'Invalid mobile numbers (34)')
-    cy.get('.moj-sub-navigation__link').last().should('contain.text', 'Missing mobile numbers (1,876)')
+    cy.get('.moj-sub-navigation__link').eq(0).should('contain.text', 'Invalid mobile numbers (34)')
+    cy.get('.moj-sub-navigation__link').eq(1).should('contain.text', 'Missing mobile numbers')
+    cy.get('.moj-sub-navigation__link').eq(2).should('contain.text', 'Duplicate mobile numbers')
   })
 
   it('can sort table', () => {
