+++ server/data/deliusClient.ts
@@ -1,19 +1,36 @@
 import { asSystem, RestClient } from '@ministryofjustice/hmpps-rest-client'
 import { AuthenticationClient } from '@ministryofjustice/hmpps-auth-clients'
+import superagent from 'superagent'
 import config from '../config'
 import logger from '../../logger'
+import notifyClients from './notifyClients'
 
 export default class DeliusClient extends RestClient {
   constructor(authenticationClient: AuthenticationClient) {
     super('Delius Client', config.apis.delius, logger, authenticationClient)
   }
 
-  async getUserAccess(username: string): Promise<{ providers: Provider[] }> {
+  async getUserDatasets(username: string): Promise<{ providers: Provider[] }> {
     return super.get({ path: `/users/${username}/providers` }, asSystem())
   }
 
-  async getDataQualityStats(providerCode: string): Promise<DataQualityStats> {
-    return super.get({ path: `/data-quality/${providerCode}/stats` }, asSystem())
+  async getEnabledUserProviders(username: string): Promise<Provider[]> {
+    const response = await this.getUserDatasets(username)
+    const enabledProviders = Object.keys(notifyClients)
+    return response.providers.filter(provider => enabledProviders.includes(provider.code))
+  }
+
+  async getUserAccess(username: string, crns: string[]): Promise<UserAccess> {
+    return super.post({ path: `/users/${username}/access`, data: crns }, asSystem())
+  }
+
+  async getInvalidMobileNumberCount(providerCode: string): Promise<string> {
+    return (
+      await super.get<superagent.Response>(
+        { path: `/data-quality/${providerCode}/invalid-mobile-numbers/count`, raw: true },
+        asSystem(),
+      )
+    ).text
   }
 
   async getInvalidMobileNumbers(providerCode: string, page: number, sort?: string): Promise<Page<DataQualityEntry>> {
@@ -36,14 +53,24 @@ export default class DeliusClient extends RestClient {
     )
   }
 
-  mapSort(sortStr?: string): string {
-    let sort = ''
+  async getDuplicateMobileNumbers(providerCode: string, page: number, sort?: string): Promise<Page<DataQualityEntry>> {
+    return super.get(
+      {
+        path: `/data-quality/${providerCode}/duplicate-mobile-numbers`,
+        query: `page=${page - 1}${this.mapSort(sort, `&sort=mobileNumber,desc`)}`,
+      },
+      asSystem(),
+    )
+  }
+
+  mapSort(sortStr?: string, defaultSort: string = '&sort=forename,asc&sort=surname,asc'): string {
+    let sort = defaultSort
     if (sortStr) {
       const sortOptions = sortStr.split('.')
       const field = sortOptions[0]
       const direction = sortOptions.length > 1 && sortOptions[1] === 'descending' ? 'desc' : 'asc'
 
-      if (!field || field === 'Name') sort = `&sort=forename,${direction}&sort=surname,${direction}`
+      if (field === 'Name') sort = `&sort=forename,${direction}&sort=surname,${direction}`
       if (field === 'CRN') sort = `&sort=crn,${direction}`
       if (field === 'Mobile number') sort = `&sort=mobileNumber,${direction}`
       if (field === 'Manager')
@@ -58,11 +85,6 @@ export default class DeliusClient extends RestClient {
 export interface Provider {
   code: string
   name: string
-}
-
-export interface DataQualityStats {
-  invalid: number
-  missing: number
 }
 
 export interface DataQualityEntry {
@@ -84,4 +106,16 @@ export interface Page<T> {
     totalPages: number
     totalElements: number
   }
+}
+
+export interface CaseAccess {
+  crn: string
+  userExcluded: boolean
+  userRestricted: boolean
+  exclusionMessage?: string
+  restrictionMessage?: string
+}
+
+export interface UserAccess {
+  access: CaseAccess[]
 }
